+++ src/main/java/org/example/playus/domain/employee/EmployeeController.java
@@ -1,11 +1,13 @@
 package org.example.playus.domain.employee;
 
 import lombok.RequiredArgsConstructor;
+import org.example.playus.domain.employee.dto.EmployeeExpDetailResponseDto;
+import org.example.playus.domain.employee.dto.EmployeeExpReponseDto;
+import org.example.playus.domain.employee.model.Employee;
 import org.example.playus.domain.security.service.UserDetailsImpl;
 import org.example.playus.global.common.CommonResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
