+++ src/main/java/org/example/playus/domain/employee/EmployeeRepositoryMongo.java
@@ -1,5 +1,6 @@
 package org.example.playus.domain.employee;
 
+import org.example.playus.domain.employee.model.Employee;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
