+++ src/main/java/org/example/playus/domain/employee/model/Account.java
@@ -1,4 +1,4 @@
-package org.example.playus.domain.employee;
+package org.example.playus.domain.employee.model;
 
 import lombok.Builder;
 import lombok.Getter;
