+++ src/main/java/org/example/playus/domain/employee/model/PersonalInfo.java
@@ -1,4 +1,4 @@
-package org.example.playus.domain.employee;
+package org.example.playus.domain.employee.model;
 
 import lombok.Getter;
