--- src/main/java/org/example/playus/domain/quest/groupGuset/WeeklyInfoRepositoryMongo.java
@@ -1,8 +1,0 @@
-package org.example.playus.domain.quest.groupGuset;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WeeklyInfoRepositoryMongo extends MongoRepository<WeeklyInfo, String> {
-}
