+++ src/main/java/org/example/playus/domain/security/filter/JwtAuthenticationFilter.java
@@ -7,9 +7,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.example.playus.domain.admin.Role;
-import org.example.playus.domain.employee.Employee;
+import org.example.playus.domain.employee.model.Employee;
 import org.example.playus.domain.employee.EmployeeRepositoryMongo;
-import org.example.playus.domain.employee.TokenStore;
+import org.example.playus.domain.employee.model.TokenStore;
 import org.example.playus.domain.login.dto.LoginRequestDto;
 import org.example.playus.domain.security.jwt.JwtUtil;
 import org.example.playus.global.exception.CustomException;
