+++ src/main/java/org/example/playus/domain/security/service/UserDetailsServiceImpl.java
@@ -1,6 +1,6 @@
 package org.example.playus.domain.security.service;
 
-import org.example.playus.domain.employee.Employee;
+import org.example.playus.domain.employee.model.Employee;
 import org.example.playus.domain.employee.EmployeeRepositoryMongo;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
