+++ src/main/java/org/example/playus/domain/signup/SignupService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.playus.domain.admin.Admin;
 import org.example.playus.domain.admin.Role;
-import org.example.playus.domain.employee.Account;
-import org.example.playus.domain.employee.Employee;
+import org.example.playus.domain.employee.model.Account;
+import org.example.playus.domain.employee.model.Employee;
 import org.example.playus.domain.employee.EmployeeRepositoryMongo;
 import org.example.playus.domain.sheet.GoogleSheetsHelper;
 import org.example.playus.domain.signup.dto.SignupRequestDto;
@@ -22,7 +22,6 @@ import java.util.Arrays;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
